Reorganize into validation subpackge
