+++ ballet/util/ci.py
@@ -8,12 +8,6 @@ from ballet.exc import UnexpectedTravisEnvironmentError
 from ballet.util.git import PullRequestBuildDiffer
 from ballet.util.log import logger
 
-TEST_TYPE_ENV_VAR = 'TEST_TYPE'
-
-
-def detect_target_type():
-    return os.environ.get(TEST_TYPE_ENV_VAR, default=None)
-
 
 def get_travis_env_or_fail(name):
     if name in os.environ:
@@ -38,7 +32,7 @@ def dump_travis_env_vars():
 
 def get_travis_pr_num():
     """Return the PR number if the job is a pull request, None otherwise
-    
+
     Returns:
         int
