+++ ballet/validation/osfs.py
@@ -1,0 +1,23 @@
+from ballet.validation.base import (
+    PostAcceptanceFeaturePerformanceEvaluator,
+    PreAcceptanceFeaturePerformanceEvaluator)
+
+
+class FeatureRelevanceEvaluator(PreAcceptanceFeaturePerformanceEvaluator):
+    """Accept a feature if it is correlated to the target"""
+
+    def judge(self, feature):
+        return False
+
+
+class FeatureRedundancyEvaluator(PostAcceptanceFeaturePerformanceEvaluator):
+    """Remove a feature if it is conditionally independent of the target
+
+    Let Sk be the set of subsets of features of size less than or equal to k.
+    A feature Xi is redundant if it is independent of the target, conditional
+    on some S in Sk. If a feature is redundant, it is removed from the feature
+    matrix.
+    """
+
+    def prune(self, k=4):
+        return []
